dlp-gateway/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
redis[asyncio]>=5.0
rapidfuzz>=3.0
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0
pyyaml>=6.0
# gliner>=0.1.0  # optional
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pdfplumber>=0.10.0
python-docx>=1.1.0
aiohttp>=3.9.0

# torch==2.4.0+cpu  # optional — only needed if gliner NER is enabled
